requirements.txt
absl-py==0.12.0
APScheduler==3.6.3
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dill==0.3.3
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
importlib-resources==5.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.3
Pillow==8.1.2
promise==2.3
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
telegram-bot @ file:///Users/fkrach/Code/GitRepos/Telegram-Bot/dist/telegram_bot-0.0.1-py3-none-any.whl
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
tornado==6.1
tqdm==4.59.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
